UseBasicParsing
Running against certain machines where IE Configuration has not occurred, the script generates an error:
Invoke-WebRequest : The response content cannot be parsed because the Internet Explorer engine is not available, or 
Internet Explorer's first-launch configuration is not complete. Specify the UseBasicParsing parameter and try again. 

UseBasicParsing seems to work around this issue.